+++ packages/builder/src/components/backend/DataTable/buttons/RowActionsButton.svelte
@@ -1,16 +1,9 @@
 <script>
-  import {
-    ActionButton,
-    Body,
-    Button,
-    Drawer,
-    DrawerContent,
-    Table,
-  } from "@budibase/bbui"
+  import { ActionButton } from "@budibase/bbui"
+  import RowActionsDrawer from "components/common/rowActions/RowActionsDrawer.svelte"
 
   export let tableId
   let drawer
-  let upsertDrawer
 
   let rowActions = []
 
@@ -24,42 +17,11 @@
     drawer.show()
   }
 
-  function addNewAction() {
-    upsertDrawer.show()
-  }
-
   $: hasRowActions = rowActions > 0
   $: title = "Row actions" + (hasRowActions ? ` (${rowActions.length})` : "")
-
-  function closeAllDrawers() {
-    upsertDrawer.hide()
-  }
 </script>
 
 <ActionButton icon="JourneyAction" quiet on:click={openDrawer}>
   {title}
 </ActionButton>
-<Drawer
-  bind:this={drawer}
-  title={"Row actions"}
-  forceModal
-  on:drawerHide={closeAllDrawers}
->
-  <div slot="buttons">
-    <Button cta on:click={addNewAction}>Add new</Button>
-  </div>
-  <DrawerContent
-    slot="body"
-    title="Row actions"
-    showConfirmButton={false}
-    showCancelButton={false}
-  >
-    {#if !rowActions.length}
-      <Body size="S">No row actions are created for this table.</Body>
-    {:else}
-      <Table />
-    {/if}
-  </DrawerContent>
-</Drawer>
-
-<Drawer bind:this={upsertDrawer} title={"Add new row action"} />
+<RowActionsDrawer bind:drawer {rowActions} />
